requirements.txt
notebook
nbgrader

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

tqdm
pandas

scipy
numpy
matplotlib
